threaded crawler: move code at better place
