+++ monocle/main.py
@@ -26,99 +26,17 @@ import sys
 import os
 import yaml
 
-from time import sleep
-from datetime import datetime
 from pprint import pprint
 
 from monocle import utils
 from monocle.db.db import ELmonocleDB
-from monocle.github.graphql import GithubGraphQLQuery
 from monocle.github import pullrequest
 from monocle.gerrit import review
 from monocle.envdefault import EnvDefault
+from monocle.crawler import Crawler
 
-from threading import Thread
-from dataclasses import dataclass
 from monocle import projects
 from jsonschema import validate
-
-
-@dataclass
-class GithubCrawlerArgs(object):
-    updated_since: str
-    loop_delay: int
-    command: str
-    org: str
-    base_url: str
-    token: str
-
-
-@dataclass
-class GerritCrawlerArgs(object):
-    updated_since: str
-    loop_delay: int
-    command: str
-    base_url: str
-    repository: str
-
-
-class MonocleCrawler(Thread):
-
-    log = logging.getLogger("monocle.Crawler")
-
-    def __init__(self, args):
-        super().__init__()
-        self.updated_since = args.updated_since
-        self.loop_delay = int(args.loop_delay)
-        self.get_one = getattr(args, 'id', None)
-        self.db = ELmonocleDB(args.elastic_conn)
-        if args.command == 'github_crawler':
-            self.get_one_rep = getattr(args, 'repository', None)
-            self.org = args.org
-            self.repository_el_re = args.org.lstrip('^') + '.*'
-            self.prf = pullrequest.PRsFetcher(
-                GithubGraphQLQuery(args.token),
-                args.base_url, args.org)
-        elif args.command == 'gerrit_crawler':
-            self.repository_el_re = args.repository.lstrip('^')
-            self.prf = review.ReviewesFetcher(
-                args.base_url, args.repository)
-
-    def get_last_updated_date(self):
-        change = self.db.get_last_updated(self.repository_el_re)
-        if not change:
-            return (
-                self.updated_since or
-                datetime.now().strftime("%Y-%m-%dT%H:%M:%SZ"))
-        else:
-            logging.info(
-                "Most recent change date in the database for %s is %s" % (
-                    self.repository_el_re, change['updated_at']))
-            return change['updated_at']
-
-    def run_step(self):
-        updated_since = self.get_last_updated_date()
-        prs = self.prf.get(updated_since)
-        objects = self.prf.extract_objects(prs)
-        if objects:
-            self.log.info("%s objects will be updated in the database" % len(
-                objects))
-            self.db.update(objects)
-
-    def run(self):
-        if self.get_one:
-            if not self.get_one_rep:
-                print("The --repository argument must be given")
-            else:
-                pprint(self.prf.get_one(
-                    self.org, self.get_one_rep,
-                    self.get_one))
-        else:
-            while True:
-                self.run_step()
-                self.log.info("Waiting %s seconds before next fetch ..." % (
-                    self.loop_delay))
-                sleep(self.loop_delay)
 
 
 def main():
@@ -214,7 +132,7 @@ def main():
             print('%s does not start with http:// or https://' % args.base_url,
                   file=sys.stderr)
             return 1
-        crawler = MonocleCrawler(args)
+        crawler = Crawler(args)
         crawler.start()
 
     if args.command == "crawler":
@@ -229,23 +147,23 @@ def main():
         for project in config['projects']:
             for crawler_item in project['crawler'].get(
                     'github_orgs', []):
-                args = GithubCrawlerArgs(
+                args = pullrequest.GithubCrawlerArgs(
                     command='github_crawler',
                     org=crawler_item['name'],
                     updated_since=crawler_item['updated_since'],
                     loop_delay=project['crawler']['loop_delay'],
                     token=crawler_item['token'],
                     base_url=crawler_item['base_url'])
-                tpool.append(MonocleCrawler(args))
+                tpool.append(Crawler(args))
             for crawler_item in project['crawler'].get(
                     'gerrit_repositories', []):
-                args = GerritCrawlerArgs(
+                args = review.GerritCrawlerArgs(
                     command='gerrit_crawler',
                     repository=crawler_item['name'],
                     updated_since=crawler_item['updated_since'],
                     loop_delay=project['crawler']['loop_delay'],
                     base_url=crawler_item['base_url'])
-                tpool.append(MonocleCrawler(args))
+                tpool.append(Crawler(args))
         for cthread in tpool:
             cthread.start()
